Implement User sign up
Implemented User creation by creating the respective route in UsersController and business logic in UsersService controller. CreateUserDto was created in order to specify required data in the request body.

Entities were also explicitly named. Specially, the User entity was renamed to UserEntity because PostgreSQL does not allow a table named user, which was causing an error.

Finally PersistenceConfig was added to enable JPA auditing. AuditingEntityListener was added to all entities to make this work.