Implement GET my_posts endpoint
Renamed AdoptionApplicationDto to AdoptionApplicationInfoDto and updated its respective documentation to convey a more specific purpose: represent a limited view of the data.

Implemented GET pets/my_posts endpoint in a similar way to GET adoptions/my_applications endpoint. Namely, implemented the route in PetsController, business logic in PetsService and declared a method to retrieve the desired information in PetsRepository. A PetInfoDto was also implemented to represent a limited view of the data.